Add comments to code for readability. moved resolvers/typeDefs into schemas folder. Adjusted imports.
