+++ server/src/models/index.ts
@@ -2,8 +2,8 @@ import Book from './Book.js';
 import User from './User.js';
 import Post from './Post.js';
 import Comment from './Comment.js';
-import typeDefs from './typeDefs.js';
-import resolvers from './resolvers.js';
+import typeDefs from '../schemas/typeDefs.js';
+import resolvers from '../schemas/resolvers.js';
 
 export { Book, User, Post, Comment };
 export { typeDefs, resolvers };
